CARDS-1780: Configure/Implement data import from Clarity
Refactor code moving it into the clarity-integration core CARDS module
