+++ distribution/src/main/features/core/cards.json
@@ -72,10 +72,6 @@
         {
             "id":"io.uhndata.cards:cards-modules-vocabularies-ignore:${project.version}",
             "start-order":"25"
-        },
-        {
-            "id":"io.uhndata.cards:cards-clarityimport:${project.version}",
-            "start-order":"25"
         }
     ]
 }
