+++ modules/clarity-integration/src/main/features/feature.json
@@ -18,6 +18,10 @@
 {
     "bundles":[
       {
+        "id":"org.osgi:org.osgi.service.jdbc:1.0.1",
+        "start-order":"25"
+      },
+      {
         "id":"com.microsoft.sqlserver:mssql-jdbc:11.2.1.jre11",
         "start-order":"25"
       },
