+++ modules/clarity-integration/src/main/resources/SLING-INF/content/mapping.xml
