+++ modules/clarity-integration/src/main/resources/SLING-INF/nodetypes/mapping.cnd
