+++ modules/pom.xml
@@ -67,5 +67,6 @@
     <module>webhook-backup</module>
     <module>saml</module>
     <module>scheduled-csv-export</module>
+    <module>clarity-integration</module>
   </modules>
 </project>
