+++ prems-resources/pom.xml
@@ -34,6 +34,5 @@
   <modules>
     <module>clinical-data</module>
     <module>feature</module>
-    <module>clarity-import</module>
   </modules>
 </project>
